+++ lib/features/appointment/view/appointment_page.dart
@@ -2,8 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:go_router/go_router.dart';
 import 'package:meddly/core/core.dart';
-import 'package:meddly/features/appointment/view/view.dart';
-import 'package:meddly/features/appointment/widgets/appointment_body.dart';
+import 'package:meddly/features/appointment/appointment.dart';
 import 'package:meddly/features/browse/browse.dart';
 import 'package:meddly/l10n/l10n.dart';
 
@@ -36,36 +35,6 @@ class AppointmentPage extends HookWidget {
         child: const Icon(Icons.add),
       ),
       body: AppointmentBody(controller),
-    );
-  }
-}
-
-class AppointmentTabBar extends HookWidget {
-  const AppointmentTabBar(
-    this.controller, {
-    super.key,
-  });
-
-  final TabController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return TabBar(
-      labelStyle: context.textTheme.titleSmall,
-      unselectedLabelStyle: context.textTheme.titleSmall!.copyWith(
-        fontWeight: FontWeight.w400,
-      ),
-      labelColor: context.colorScheme.onBackground,
-      indicatorColor: context.colorScheme.primary,
-      controller: controller,
-      tabs: [
-        Tab(
-          text: context.l10n.upcomingAppointments,
-        ),
-        Tab(
-          text: context.l10n.pastAppointments,
-        ),
-      ],
     );
   }
 }
